Allow force field parameters to be changed for individual atoms
Force field parameters are obtained by a call to
ChemicalObject.getAtomProperty, which starts lookup at the top-level
chemical object, thus permitting chemical object types to override
paramters defined for its constituent objects (atoms and groups). With
the current change, an even higher priority is given to atom
properties defined as attributes of the individual atom object (as
opposed to its atom type object). This makes it possible to override
force field parameters by a simple assignment in a Python script.